requirements.txt
fastapi~=0.128.0
sqlmodel
pydantic-settings
psycopg2-binary
asyncpg==0.29.0
greenlet
